Revert feedback-tab
