+++ core/templates/pages/exploration-editor-page/exploration-editor-page.component.ts
@@ -54,7 +54,7 @@ require(
 require(
   'pages/exploration-editor-page/editor-tab/' +
   'exploration-editor-tab.component.ts');
-require('pages/exploration-editor-page/feedback-tab/feedback-tab.component.ts');
+require('pages/exploration-editor-page/feedback-tab/feedback-tab.directive.ts');
 require(
   'pages/exploration-editor-page/feedback-tab/thread-table/' +
   'thread-table.component.ts');
